+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Added
+- Allow searching case-insensitively via `-i`/`--ignore-case` option
+
 ## 0.2.5 - 2020-06-05
 ### Added
 - Allow filtering results to a specific path (directory or file) via `-p`/`--path` option